Run tests in browser
Runs relevant tests in browser using puppeteer package. Results
are captured in a snapshot with duration information hidden and
compared against an image of the last known passing image run. This
is a bit brittle in that it requires manually updating the current
passing screenshot when browser tests are modified.

Resolves Issue: https://github.com/haadcode/logplease/issues/8